Fix local training
